Feature/audio (#37)
* [MOBILE] audio manager

* [BACK] settings volume

* [MOBILE] connect settings with audio manager

* [MOBILE] refactor settings loading

+++ mobile/app/src/main/java/com/bellako/kiwi/features/settings/SettingsDTO.kt
@@ -4,8 +4,8 @@ import com.bellako.kiwi.types.Email
 
 data class SettingsDTO(
     val email: String,
-    val soundVolume: Int,
-    val musicVolume: Int
+    val soundVolume: Float,
+    val musicVolume: Float
 ) {
     fun toState(): SettingsState {
         return SettingsState(
